techpack: camera: Eliminate float point usage
Kernel drivers do not use floating point as it cannot easily
trap itself, unlike user-space where the kernel catches a trap
and then initiates the transitions from integer to floating point
mode.
Floating point usage in kernel requires saving and restoration of
floating point registers, with many other possible work.
Eliminate floating point to attain parity with the rest of the
kernel-space as these drivers are built into the kernel.

+++ drivers/media/platform/msm/camera_8150/cam_sensor_module/cam_ois/onsemi_fw/LC898124/Accuracy.h
@@ -8,20 +8,20 @@
 //#define		RADIUS 		75							// 75um
 
 // Parameter define
-//#define		DEGSTEP		3							// Degree of one step (3Åã)
-//#define		ACCURACY	3.0F						// Accuracy (Å}3.0um)
+//#define		DEGSTEP		3							// Degree of one step (3ÔøΩÔøΩ)
+//#define		ACCURACY	3.0F						// Accuracy (ÔøΩ}3.0um)
 //#define		WAIT_MSEC	10							// Each step wait time(msec)
 #define		LOOPTIME	3							// Read times at each step
 
 // Constants
-#define		PI			3.14159		// ÉŒ
+#define		PI			3.14159		// ÔøΩÔøΩ
 
 
 typedef struct tag_Dual_Axis
 {
-	float xpos;
-	float xhall;
-	float ypos;
-	float yhall;
+	int xpos;
+	int xhall;
+	int ypos;
+	int yhall;
 }Dual_Axis_t;
