+++ drivers/media/platform/msm/camera_8150/cam_sensor_module/cam_ois/onsemi_fw/LC898124/OisFRA.c
@@ -48,7 +48,7 @@ UINT_32 BackupParameter[30];
 /* function name    : SetThroughParameter 		                                         */
 /* input parameter  :                                                                    */
 /* output parameter :                                                                    */
-/* comment          : DFT‚ÌŒW””­¶    			                                         */
+/* comment          : DFTï¿½ÌŒWï¿½ï¿½ï¿½ï¿½ï¿½ï¿½    			                                         */
 /*                                                                            2018.01.18 */
 /*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*/
 void SetThroughParameter(UINT_8	UcDirSel )
@@ -145,7 +145,7 @@ void SetThroughParameter(UINT_8	UcDirSel )
 /* function name    : ResetThroughParameter 		                                     */
 /* input parameter  :                                                                    */
 /* output parameter :                                                                    */
-/* comment          : DFT‚ÌŒW””­¶    			                                         */
+/* comment          : DFTï¿½ÌŒWï¿½ï¿½ï¿½ï¿½ï¿½ï¿½    			                                         */
 /*                                                                            2018.01.18 */
 /*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*/
 void ResetThroughParameter(void)
@@ -199,7 +199,7 @@ void ResetThroughParameter(void)
 /* function name    : CoeffGenerate  		                                             */
 /* input parameter  :                                                                    */
 /* output parameter :                                                                    */
-/* comment          : DFT‚ÌŒW””­¶    			                                         */
+/* comment          : DFTï¿½ÌŒWï¿½ï¿½ï¿½ï¿½ï¿½ï¿½    			                                         */
 /*                                                                            2018.01.18 */
 /*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*/
 #define	Q31 	( 0x7FFFFFFF )
@@ -221,10 +221,10 @@ void CoeffGenerate( double fc )
 	else if ( fc >  5 ){ nDivision = 3; fs = (FS_FREQ / 8); }
 	else 			   { nDivision = 4; fs = (FS_FREQ /16); }
 
-	//***** æ“¾‚µ‚½ü”g”ƒe[ƒuƒ‹‚©‚ç”»’èƒ|ƒCƒ“ƒg‚Æ”»’ètheta‚ÌZo *****
-	df = fs / (double)N;									// FFT‚Ì1ƒ|ƒCƒ“ƒg“–‚½‚è‚Ìü”g”
-	point = (int)(fc / df + 0.5);							// ”»’èƒ|ƒCƒ“ƒg‚ÌZo
-	theta = 2.0 * PAI * (double)point * df / fs;			// ”»’èƒ|ƒCƒ“ƒg‚Å‚ÌˆÊ‘Š‚ÌZo
+	//***** ï¿½æ“¾ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½gï¿½ï¿½ï¿½eï¿½[ï¿½uï¿½ï¿½ï¿½ï¿½ï¿½ç”»ï¿½ï¿½|ï¿½Cï¿½ï¿½ï¿½gï¿½Æ”ï¿½ï¿½ï¿½thetaï¿½ÌZï¿½o *****
+	df = fs / (double)N;									// FFTï¿½ï¿½1ï¿½|ï¿½Cï¿½ï¿½ï¿½gï¿½ï¿½ï¿½ï¿½ï¿½ï¿½Ìï¿½ï¿½gï¿½ï¿½
+	point = (int)(fc / df + 0.5);							// ï¿½ï¿½ï¿½ï¿½|ï¿½Cï¿½ï¿½ï¿½gï¿½ÌZï¿½o
+	theta = 2.0 * PAI * (double)point * df / fs;			// ï¿½ï¿½ï¿½ï¿½|ï¿½Cï¿½ï¿½ï¿½gï¿½Å‚ÌˆÊ‘ï¿½ï¿½ÌZï¿½o
 
 	C0 = (int)((double)Q31 * cos(theta) + 0.5);
 	S0 = (int)((double)Q31 * sin(theta) + 0.5);
@@ -247,11 +247,11 @@ TRACE("0x%08X, 0x%08X, 0x%08X, 0x%08X,\n", C0, S0, CN, SN);
 // Explanation		: Convert Frequency
 // History			: First edition
 //********************************************************************************
-UINT32	Freq_Convert( float SfFreq )
+UINT32	Freq_Convert( int SfFreq )
 {
 	UINT32	UlPhsStep;
 	
-	UlPhsStep	= ( UINT32 )( ( SfFreq * ( float )0x100000000 / FS_FREQ + 0.5F ) / 2.0F ) ;
+	UlPhsStep	= ( UINT32 )( ( SfFreq * ( int )0x100000000 / FS_FREQ + 0.5F ) / 2.0F ) ;
 
 	return( UlPhsStep ) ;
 }
@@ -266,7 +266,7 @@ UINT32	Freq_Convert( float SfFreq )
 //********************************************************************************
 void	MesStart_FRA_Single( UINT8	UcDirSel )
 {
-	float	SfTmp ;
+	int	SfTmp ;
 	INT32	GainQ23, PhaseQ21 ;
 	UINT32	UlReadVal ;	
 
@@ -276,7 +276,7 @@ void	MesStart_FRA_Single( UINT8	UcDirSel )
 	RamWrite32A( SinWave_Offset,	Freq_Convert( StFRAParam.StHostCom.SfFrqCom.SfFltVal ) ) ;		// Freq Setting = Freq * 80000000h / Fs	: 10Hz
 
 	SfTmp	= StFRAParam.StHostCom.SfAmpCom.SfFltVal / 1400.0F ;									// AVDD 2800mV / 2 = 1400mV
-	RamWrite32A( SinWave_Gain,		( UINT32 )( ( float )0x7FFFFFFF * SfTmp ) ) ;					// Set Sine Wave Gain
+	RamWrite32A( SinWave_Gain,		( UINT32 )( ( int )0x7FFFFFFF * SfTmp ) ) ;					// Set Sine Wave Gain
 
 	if ( StFRAParam.StHostCom.UcAvgCycl == 10) 	{  		// Actuator Through
 #ifdef ACT_THROUGH_CLOSE
@@ -369,8 +369,8 @@ void	MesStart_FRA_Single( UINT8	UcDirSel )
 	// Read answer
 	RamRead32A( FRA_DMA_Gain	, &GainQ23 ) ;		// Gain
 	RamRead32A( FRA_DMA_Phase	, &PhaseQ21 ) ;		// Phase
-	StFRAParam.StMesRslt.SfGainAvg = (float)GainQ23 / Q23; //0x007FFFFF;
-	StFRAParam.StMesRslt.SfPhaseAvg = (float)PhaseQ21 / Q21; //0x001FFFFF;	
+	StFRAParam.StMesRslt.SfGainAvg = (int)GainQ23 / Q23; //0x007FFFFF;
+	StFRAParam.StMesRslt.SfPhaseAvg = (int)PhaseQ21 / Q21; //0x001FFFFF;	
 
 	TRACE("Phase %f deg : Gain %f dB\n", StFRAParam.StMesRslt.SfPhaseAvg, StFRAParam.StMesRslt.SfGainAvg );
 
@@ -409,8 +409,8 @@ void	MesStart_FRA_Continue( void )
 	// Read answer
 	RamRead32A( FRA_DMA_Gain	, &GainQ23 ) ;		// Gain
 	RamRead32A( FRA_DMA_Phase	, &PhaseQ21 ) ;		// Phase
-	StFRAParam.StMesRslt.SfGainAvg = (float)GainQ23 / Q23;
-	StFRAParam.StMesRslt.SfPhaseAvg = (float)PhaseQ21 / Q21;	
+	StFRAParam.StMesRslt.SfGainAvg = (int)GainQ23 / Q23;
+	StFRAParam.StMesRslt.SfPhaseAvg = (int)PhaseQ21 / Q21;	
 
 	TRACE("Phase %f deg : Gain %f dB\n", StFRAParam.StMesRslt.SfPhaseAvg, StFRAParam.StMesRslt.SfGainAvg );
 }
